backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pypandoc_binary